+++ keymaster/include/optee_keymaster/ipc/optee_keymaster_ipc.h
@@ -1,5 +1,5 @@
 /*
- *
+ * Copyright (C) 2015 The Android Open Source Project
  * Copyright (C) 2017 GlobalLogic
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -14,29 +14,28 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
- #ifndef OPTEE_KEYMASTER_IPC_H
- #define OPTEE_KEYMASTER_IPC_H
+
+#ifndef OPTEE_KEYMASTER_OPTEE_KEYMASTER_IPC_H_
+#define OPTEE_KEYMASTER_OPTEE_KEYMASTER_IPC_H_
 
 #include <keymaster/android_keymaster_messages.h>
 #include <optee_keymaster/ipc/keymaster_ipc.h>
+
 __BEGIN_DECLS
 
-#define TA_KEYMASTER_UUID { 0xdba51a17, 0x0563, 0x11e7, \
-		{ 0x93, 0xb1, 0x6f, 0xa7, 0xb0, 0x07, 0x1a, 0x51} }
-
 const uint32_t OPTEE_KEYMASTER_RECV_BUF_SIZE = 2 * PAGE_SIZE;
-const uint32_t OPTEE_KEYMASTER_SEND_BUF_SIZE = 2 * PAGE_SIZE;
+const uint32_t OPTEE_KEYMASTER_SEND_BUF_SIZE =
+        (PAGE_SIZE - sizeof(struct keymaster_message) - 16 /* tipc header */);
 
-int optee_keymaster_initialize(void);
 int optee_keymaster_connect(void);
-
-keymaster_error_t optee_keymaster_call(uint32_t cmd, const keymaster::Serializable& req,
-                        keymaster::KeymasterResponse* rsp);
-
+int optee_keymaster_call(uint32_t cmd, void* in, uint32_t in_size, uint8_t* out,
+                          uint32_t* out_size);
 void optee_keymaster_disconnect(void);
-void optee_keymaster_finalize(void);
 
-const char* print_error_message(uint32_t error);
+keymaster_error_t translate_error(int err);
+keymaster_error_t optee_keymaster_send(uint32_t command, const keymaster::Serializable& req,
+                                        keymaster::KeymasterResponse* rsp);
 
 __END_DECLS
-#endif /* OPTEE_KEYMASTER_IPC_H */
+
+#endif  // OPTEE_KEYMASTER_OPTEE_KEYMASTER_IPC_H_
